pyproject.toml
[tool.poetry]
name = "Shiny Happiness"
version = "0.0.0"
description = "Evaluation of Generative (Molecule) Models"
authors = ["ONDRA-IDA-TEAM"]

[[tool.poetry.source]]
name = "pypi_"
url = "https://pypi.org/simple/"

[tool.poetry.dependencies]
python = "~=3.8.0"
smart-open = "~=5.0"
pandas = "~=1.1"
nltk = "~=3.6"
matplotlib = "~=3.4"
scikit-learn = "~=0.24"
mlflow = "^2"
typer = "~=0.4"
boto3 = "~=1.20"
tqdm = "~=4.62"
jupyterlab = "~=3.1"
ipywidgets = "~=7.6"
codetiming = "~=1.3"
papermill = "~=2.3"
streamlit = "~=1.9"
tensorboard = "~=2.9"
black = "^22.3"
pytest = "^7.1"
traitlets = "<=5.1.1"
problog = "^2.2.2"
rdkit-pypi = "^2022.3.3"
icecream = "^2.1.2"
seaborn = "^0.11.2"
pyswip = "^0.2.10"
networkx = "^2.8.6"
latex = "^0.7.0"
joblib = "^1.2.0"
chemspipy = "^2.0.0"
pyyaml = "^6.0"
imageio = "^2.28.1"
scipy = "^1.10.1"
wandb = "^0.15.2"
hydra-core = "^1.3.2"
ipdb = "^0.13.13"
pdbpp = "^0.10.3"
overrides = "^7.3.1"
pygsp = "^0.5.1"
pyemd = "^1.0.0"
setuptools = "<70"
dagshub = "^0.3.37"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
